+++ src/id_translation/dio/_dict.py
@@ -8,25 +8,24 @@ from ..types import IdType, NameType, SourceType
 from ._data_structure_io import DataStructureIO
 
 
-class DictIO(DataStructureIO):
+class DictIO(DataStructureIO[dict[NameType, IdType], NameType, SourceType, IdType]):
     """Implementation for dicts."""
 
-    # TODO types here all mostly wrong. Would require a lot of overloads to fix..
-
-    @staticmethod
-    def handles_type(arg: Any) -> bool:
+    @classmethod
+    def handles_type(cls, arg: Any) -> bool:
         return isinstance(arg, dict)
 
-    @staticmethod
-    def names(translatable: dict[NameType, IdType]) -> list[NameType]:
+    @classmethod
+    def names(cls, translatable: dict[NameType, IdType]) -> list[NameType]:
         return list(translatable)
 
-    @staticmethod
-    def extract(translatable: dict[NameType, IdType], names: list[NameType]) -> dict[NameType, Sequence[IdType]]:
+    @classmethod
+    def extract(cls, translatable: dict[NameType, IdType], names: list[NameType]) -> dict[NameType, Sequence[IdType]]:
         return {name: as_list(translatable[name]) for name in names}
 
-    @staticmethod
+    @classmethod
     def insert(
+        cls,
         translatable: dict[NameType, IdType] | dict[NameType, Sequence[IdType]],
         names: list[NameType],
         tmap: TranslationMap[NameType, SourceType, IdType],
@@ -38,10 +37,13 @@ class DictIO(DataStructureIO):
         from ._resolve import resolve_io
 
         with disable_temporarily(RESOLVE_IO_LOGGER):
-            translated = {
-                key: resolve_io(value).insert(value, [key], tmap, copy=True) if key in names else value
-                for key, value in translatable.items()
-            }
+            translated = {}
+            for key, value in translatable.items():
+                dio: DataStructureIO[IdType | Sequence[IdType], NameType, SourceType, IdType] = resolve_io(value)
+                if key in names:
+                    translated[key] = dio.insert(value, [key], tmap, copy=True)
+                else:
+                    translated[key] = value
 
         if copy:
             return translated