+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added integration for [polars.DataFrame](https://docs.pola.rs/py-polars/html/reference/dataframe/index.html).
+- Added integration for [dask.DataFrame](https://docs.dask.org/en/stable/dataframe.html) and Series.
+
 ### Changed
 - Consume `[transform]`-sections in auxiliary configuration files (#231).
+- Added typehints to `dio.DataStructureIO` and other `id_translation.dio` classes and functions.
+- Added functions and methods to make creating new `DataStructureIO` implementations easier.
 
 ### Fixed
 - Verify top-level sections in auxiliary configuration files.